Separate on login response classes to their own files.
This separation allows us to use OnLoginResponse from OnSignupResponse
which greatly reduces logic duplication.
